+++ lib/browser/tag/core.js
@@ -1,0 +1,236 @@
+import styleManager from './styleManager'
+
+import {
+  __TAG_IMPL,
+  __VIRTUAL_DOM,
+  GLOBAL_MIXIN,
+  RIOT_TAG,
+  RIOT_TAG_IS
+} from './../global-variables'
+
+import {
+  isString,
+  isUndefined,
+  isObject,
+  isFunction,
+  setAttr,
+  getAttr,
+  $$,
+  each,
+  mountTo,
+  extend
+} from './../util'
+
+/**
+ * Export the Tag constructor
+ * TODO: make a better tag constructor
+ */
+// export function Tag() {}
+
+/**
+ * Create a new riot tag implementation
+ * @param   { String }   name - name/id of the new riot tag
+ * @param   { String }   tmpl - tag template
+ * @param   { String }   css - custom tag css
+ * @param   { String }   attrs - root tag attributes
+ * @param   { Function } fn - user function
+ * @returns { String } name/id of the tag just created
+ */
+export function tag(name, tmpl, css, attrs, fn) {
+  if (isFunction(attrs)) {
+    fn = attrs
+
+    if (/^[\w\-]+\s?=/.test(css)) {
+      attrs = css
+      css = ''
+    } else
+      attrs = ''
+  }
+
+  if (css) {
+    if (isFunction(css))
+      fn = css
+    else
+      styleManager.add(css)
+  }
+
+  name = name.toLowerCase()
+  __TAG_IMPL[name] = { name, tmpl, attrs, fn }
+
+  return name
+}
+
+/**
+ * Create a new riot tag implementation (for use by the compiler)
+ * @param   { String }   name - name/id of the new riot tag
+ * @param   { String }   tmpl - tag template
+ * @param   { String }   css - custom tag css
+ * @param   { String }   attrs - root tag attributes
+ * @param   { Function } fn - user function
+ * @returns { String } name/id of the tag just created
+ */
+export function tag2(name, tmpl, css, attrs, fn) {
+  if (css)
+    styleManager.add(css, name)
+
+  let exists = !!__TAG_IMPL[name]
+  __TAG_IMPL[name] = { name, tmpl, attrs, fn }
+
+  if (exists && riot.util.hotReloader)
+    riot.util.hotReloader(name)
+
+  return name
+}
+
+/**
+ * Mount a tag using a specific tag implementation
+ * @param   { String } selector - tag DOM selector
+ * @param   { String } tagName - tag implementation name
+ * @param   { Object } opts - tag logic
+ * @returns { Array } new tags instances
+ */
+export function mount(selector, tagName, opts) {
+  let tags = []
+
+  // ----- helper functions -----
+
+  function addRiotTags(arr) {
+    return arr
+        .filter(e => {
+          return !/[^-\w]/.test(e)
+        })
+        .reduce((list, e) => {
+          let name = e.trim().toLowerCase()
+          return list + `,[${RIOT_TAG_IS}="${name}"]`
+        }, '')
+  }
+
+  function selectAllTags() {
+    let keys = Object.keys(__TAG_IMPL)
+    return keys + addRiotTags(keys)
+  }
+
+  function pushTags(root) {
+    if (root.tagName) {
+      let riotTag = getAttr(root, RIOT_TAG_IS) || getAttr(root, RIOT_TAG)
+
+      // have tagName? force riot-tag to be the same
+      if (tagName && riotTag !== tagName) {
+        riotTag = tagName
+        setAttr(root, RIOT_TAG_IS, tagName)
+        setAttr(root, RIOT_TAG, tagName) // this will be removed in riot 3.0.0
+      }
+
+      let tag = mountTo(root, riotTag || root.tagName.toLowerCase(), opts)
+
+      if (tag)
+        tags.push(tag)
+    } else if (root.length) {
+      each(root, pushTags) // assume nodeList
+    }
+  }
+
+  // ----- mount code -----
+
+  // inject styles into DOM
+  styleManager.inject()
+
+  if (isObject(tagName)) {
+    opts = tagName
+    tagName = 0
+  }
+
+  let elem
+  let allTags
+
+  // crawl the DOM to find the tag
+  if (isString(selector)) {
+    if (selector === '*')
+      // select all the tags registered
+      // and also the tags found with the riot-tag attribute set
+      selector = allTags = selectAllTags()
+    else
+      // or just the ones named like the selector
+      selector += addRiotTags(selector.split(/, */))
+
+    // make sure to pass always a selector
+    // to the querySelectorAll function
+    elem = selector ? $$(selector) : []
+  }
+  else
+  // probably you have passed already a tag or a NodeList
+    elem = selector
+
+  // select all the registered and mount them inside their root elements
+  if (tagName === '*') {
+    // get all custom tags
+    tagName = allTags || selectAllTags()
+    // if the root els it's just a single tag
+    if (elem.tagName)
+      elem = $$(tagName, elem)
+    else {
+      // select all the children for all the different root elements
+      var nodeList = []
+      each(elem, _el => {
+        nodeList.push($$(tagName, _el))
+      })
+      elem = nodeList
+    }
+    // get rid of the tagName
+    tagName = 0
+  }
+
+  pushTags(elem)
+
+  return tags
+}
+
+// Create a mixin that could be globally shared across all the tags
+const mixins = {}
+const globals = mixins[GLOBAL_MIXIN] = {}
+let _id = 0
+
+/**
+ * Create/Return a mixin by its name
+ * @param   { String }  name - mixin name (global mixin if object)
+ * @param   { Object }  mix - mixin logic
+ * @param   { Boolean } g - is global?
+ * @returns { Object }  the mixin logic
+ */
+export function mixin(name, mix, g) {
+  // Unnamed global
+  if (isObject(name)) {
+    mixin(`__unnamed_${_id++}`, name, true)
+    return
+  }
+
+  var store = g ? globals : mixins
+
+  // Getter
+  if (!mix) {
+    if (isUndefined(store[name]))
+      throw new Error('Unregistered mixin: ' + name)
+
+    return store[name]
+  }
+  // Setter
+  if (isFunction(mix)) {
+    extend(mix.prototype, store[name] || {})
+    store[name] = mix
+  } else
+    store[name] = extend(store[name] || {}, mix)
+}
+
+/**
+ * Update all the tags instances created
+ * @returns { Array } all the tags instances
+ */
+export function update() {
+  return each(__VIRTUAL_DOM, tag => {
+    tag.update()
+  })
+}
+
+export function unregister(name) {
+  delete __TAG_IMPL[name]
+}
