+++ lib/browser/tag/tag.js
@@ -4,7 +4,7 @@ import NamedExpr from './named'
 import update from './update'
 import { tmpl } from 'riot-tmpl'
 import mkdom from './mkdom'
-import { mixin } from '../../riot'
+import { mixin } from './core'
 
 import {
   GLOBAL_MIXIN,
