+++ lib/browser/util.js
@@ -1,5 +1,6 @@
 import {
   __TAG_IMPL,
+  T_STRING,
   T_OBJECT,
   T_UNDEF,
   SVG_TAGS_LIST,
@@ -86,6 +87,24 @@ export function isSVGTag(name) {
  */
 export function isObject(v) {
   return v && typeof v === T_OBJECT         // typeof null is 'object'
+}
+
+/**
+ * Standart undefined check
+ * @param   { * } v - value to check
+ * @returns { Boolean } -
+ */
+export function isUndefined(v) {
+  return typeof v === T_UNDEF
+}
+
+/**
+ * Standart string check
+ * @param   { * } v - value to check
+ * @returns { Boolean } -
+ */
+export function isString(v) {
+  return typeof v === T_STRING
 }
 
 /**
