+++ package.json
@@ -54,7 +54,7 @@
   },
   "devDependencies": {
     "babel-plugin-check-es2015-constants": "^6.8.0",
-    "babel-preset-es2015": "^6.9.0",
+    "babel-preset-es2015-rollup": "^1.1.1",
     "benchmark": "^2.1.0",
     "chai": "^3.5.0",
     "cheerio": "^0.20.0",
